% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIPHERdashboard-package.R
\docType{package}
\name{SIPHERdashboard-package}
\alias{SIPHERdashboard}
\alias{SIPHERdashboard-package}
\title{SIPHERdashboard: SIPHER Synthetic Population Dashboard}
\description{
This dashboard allows you to easily explore an aggregated version of the SIPHER Synthetic Population for Individuals in Great Britain 2019-2021.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Dylan-W-Lewis/SIPHERdashboard}
  \item \url{https://www.gla.ac.uk/research/az/sipher/products/syntheticpopulation/}
  \item Report bugs at \url{https://github.com/Dylan-W-Lewis/SIPHERdashboard/issues}
}

}
\author{
\strong{Maintainer}: Dylan Lewis \email{d.lewis.1@research.gla.ac.uk}

Authors:
\itemize{
  \item Andreas Hoehn \email{andreas.hoehn@gla.ac.uk}
  \item Emma Comrie \email{emma.comrie@phs.scot}
}

}
\keyword{internal}
